requirements.txt
pyaudio
openai
ddgs
pygame
SpeechRecognition
soundfile
pypdf
python-dotenv
numpy
rich
